refactor : LottoStore로 buy 로직 이동
