+++ src/main/kotlin/next/step/lotto/controller/LottoController.kt
@@ -1,13 +1,14 @@
 package next.step.lotto.controller
 
-import next.step.lotto.domain.Lottos
+import next.step.lotto.domain.LottoNumberGenerator
+import next.step.lotto.domain.LottoStore
 import next.step.racing.view.InputView
 import next.step.racing.view.OutputView
 
 fun main() {
     runCatching {
         val payment = InputView.readPayment()
-        val lottos = Lottos.buy(payment)
+        val lottos = LottoStore.buy(payment, LottoNumberGenerator.random())
         OutputView.showLottos(lottos)
         val winningNumbers = InputView.readWinningNumbers()
         val bonusNumber = InputView.readBonusNumber()
