+++ src/main/kotlin/next/step/lotto/domain/LottoStore.kt
@@ -4,12 +4,19 @@ object LottoStore {
 
     const val LOTTO_PRICE: Int = 1000
 
-    fun buy(payment: Int): Int {
-        require(canBuy(payment)) { "로또 가격을 지불해야 합니다." }
-        return payment - LOTTO_PRICE
+    fun buy(payment: Int, algorithm: LottoNumberGenerationAlgorithm): Lottos {
+        val lottos = mutableSetOf<Lotto>()
+        var remained = payment
+        while (canBuy(remained)) {
+            val lotto: Lotto = Lotto.of(algorithm())
+            if (!lottos.contains(lotto)) {
+                remained -= LOTTO_PRICE
+                lottos.add(lotto)
+            }
+        }
+        return Lottos(lottos)
     }
 
-    fun canBuy(payment: Int): Boolean = payment >= LOTTO_PRICE
+    private fun canBuy(payment: Int): Boolean = payment >= LOTTO_PRICE
 
-    fun preview(algorithm: LottoNumberGenerationAlgorithm): Lotto = Lotto.of(algorithm())
 }
