+++ src/main/kotlin/next/step/lotto/domain/Lottos.kt
@@ -1,7 +1,7 @@
 package next.step.lotto.domain
 
 @JvmInline
-value class Lottos(val lottos: Set<Lotto>) {
+value class Lottos(private val lottos: Set<Lotto>) {
 
     fun match(winningNumbers: LottoWinningNumbers, bonusNumber: LottoNumber): LottoWinningStat =
         LottoWinningStat.of(lottos.groupingBy {
@@ -16,19 +16,6 @@ value class Lottos(val lottos: Set<Lotto>) {
     fun numbers(): Set<Set<Int>> = lottos.map { it.numbers().toSet() }.toSet()
 
     companion object {
-        fun buy(payment: Int): Lottos {
-            val lottos = mutableSetOf<Lotto>()
-            var remained = payment
-            while (LottoStore.canBuy(remained)) {
-                val lotto: Lotto = LottoStore.preview(LottoNumberGenerator.random())
-                if (!lottos.contains(lotto)) {
-                    remained = LottoStore.buy(remained)
-                    lottos.add(lotto)
-                }
-            }
-            return Lottos(lottos)
-        }
-
         fun of(lottos: Set<Lotto>): Lottos = Lottos(lottos)
     }
